my-k8s-data-app/writer/requirements.txt
mysql-connector-python
prometheus_client
Flask
